Polish sticky editing and enhance styling.
The first change made was to check if the user has left an empty input for the text fields. If this occurs, the text of the element is formatted to reflect the fact that it is empty.

Additionally, the private edit functions for the text has been refactored to remove similar event handling code to remove redundancies.

Lastly, sticky note styling was enhanced to be more visually appealing. A light mobile styling was also applied, but it is not much. For the isChecked portion of the notes, fontAwesome was applied.